pyproject.toml
[tool.poetry]
name = "electronicshopapi"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
django = "^5.1"
djangorestframework = "^3.15.2"
python-dotenv = "^1.0.1"
psycopg2-binary = "^2.9.9"
black = "^24.8.0"
django-filter = "^24.3"
djangorestframework-simplejwt = "^5.3.1"
drf-yasg = "^1.21.7"
setuptools = "^72.2.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
